fix: round decimal pts values to avoid floating-point imprecision (e.g. 99.99999999999999 → 100)
Agent-Logs-Url: https://github.com/Shooshpanius/wh40kAPI/sessions/2b02125c-6ea1-4a1c-aee9-5460d17ebdfa

+++ wh40kAPI.Server/Models/BsData/BsDataCostTierSlim.cs
@@ -4,9 +4,7 @@ namespace wh40kAPI.Server.Models.BsData;
 /// Lightweight projection of <see cref="BsDataCostTier"/> used by the
 /// <c>/unitsList</c> endpoint.  Only the fields actually consumed by the
 /// client are included; <c>Id</c> and <c>UnitId</c> are omitted to
-/// reduce payload size.  <c>Points</c> is <see langword="double"/> rather
-/// than <see langword="decimal"/> to avoid unnecessary trailing zeros in JSON
-/// serialization.
+/// reduce payload size.
 /// </summary>
 public class BsDataCostTierSlim
 {
@@ -15,5 +13,5 @@ public class BsDataCostTierSlim
     /// <summary>Maximum number of models in this cost tier (null if unbounded).</summary>
     public int? MaxModels { get; set; }
     /// <summary>Points cost for this tier.</summary>
-    public double Points { get; set; }
+    public decimal Points { get; set; }
 }
